fix: Next.js build error by separating client and server components
The build was failing on Render because `app/tools/page.tsx` was marked with 'use client' while also exporting `metadata`. This is disallowed by Next.js.

To resolve this, the component has been split into two:
- `app/tools/page.tsx` is now a server component that handles the `metadata` export.
- The client-side logic (UI and state management) has been moved to a new client component, `app/tools/client-page.tsx`.

This change aligns with Next.js best practices and resolves the build error.
